+++ src/app/captcha/serverActions/checkCaptcha.ts
@@ -9,22 +9,25 @@ const createCaptchaSuccessCookie = async() => {
 	const cookieStore = await cookies();
 	const secret = base64url.decode(process.env.JWE_SECRET_KEY as string);
 	const isProduction = process.env.RUNTIME_ENVIRONMENT === "production";
+	const maxAge = 60 * 6;
 	const encryptedSuccess = await new CompactEncrypt(new TextEncoder().encode("success"))
 	.setProtectedHeader({ alg: "dir", enc: "A256GCM" })
 	.encrypt(secret);
 
-	cookieStore.set({name: "captchaSuccess", value: encryptedSuccess, httpOnly: true, secure: isProduction, sameSite: "strict", expires: new Date(Date.now() + 1000 *60 * 6 )});
+	cookieStore.set({name: "captchaSuccess", value: encryptedSuccess, httpOnly: true, secure: isProduction, sameSite: "strict", maxAge });
 };
 
 const checkCaptcha = async(form: FormData) => {
 	const userAnswer = form.get("answer");
-	const encryptedCaptcha = form.get("encryptedCaptcha") ? form.get("encryptedCaptcha") as string: ""; 
+	const encryptedCaptcha = form.has("encryptedCaptcha") ? form.get("encryptedCaptcha") as string : ""; 
 	const captchaText = await decryptCaptcha(encryptedCaptcha);
 
 	if (captchaText === userAnswer) {
 		await createCaptchaSuccessCookie();
+
 		redirect(SiteMap.Contact);
 	} else {
+
 		redirect(SiteMap.Robot);
 	}
 };