+++ src/app/contact/components/serverActions/emailUserMessage.ts
@@ -6,6 +6,7 @@ import SMTPTransport from "nodemailer/lib/smtp-transport";
 import decryptCaptcha from "@/utils/decryptCaptcha";
 import { redirect } from "next/navigation";
 import { SiteMap } from "@/utils/siteMap";
+import { base64url, CompactEncrypt } from "jose";
 
 interface EmailForm {
 	reason: FormDataEntryValue | null;
@@ -27,7 +28,7 @@ export async function emailUserMessage(form: FormData) {
 	const name = form.get("name");
 	const email = form.get("email");
 	const message = form.get("message");
-	const isValid = await emailSuccess({reason, name, email, message});
+	const isValid = await isValidEmailForm({reason, name, email, message});
 
 	if (isValid) {
 		await transporter.sendMail({
@@ -36,19 +37,31 @@ export async function emailUserMessage(form: FormData) {
 			subject: reason,
 			text: `Name: ${name}\nFrom: ${email}\n\nMessage: ${message}` 
 		} as SMTPTransport.MailOptions);
-		redirect(SiteMap.FormProcess);
+		await createMessageSentCookie();
 
-	} else {
-	console.log("Submission failed! Try reloading the page and filling the form out faster.")
-		
 	}
+
+	redirect(SiteMap.FormProcess);
 }
 
 
-async function emailSuccess(formElements: EmailForm):Promise<boolean> {
-	const encryptedCaptcha= await cookies().then(x => x.get("captchaSuccess")?.value);
+async function isValidEmailForm(formElements: EmailForm):Promise<boolean> {
+	const encryptedCaptcha = await cookies().then(x => x.get("captchaSuccess")?.value);
 	if (encryptedCaptcha === undefined) return false;
 	const decryptedCaptcha = await decryptCaptcha(encryptedCaptcha); 
 
 	return decryptedCaptcha === "success" && formElements.reason && formElements.name && formElements.email && formElements.message ? true : false; 
+}
+
+
+async function createMessageSentCookie() {
+	const cookieStore = await cookies();
+	cookieStore.delete("captchaSuccess");
+	const secret = base64url.decode(process.env.JWE_SECRET_KEY as string);
+	const encryptedMessage = new CompactEncrypt(new TextEncoder().encode("Message Sent Successfully"))
+	.setProtectedHeader({ alg: "dir", enc: "A256GCM" })
+	.encrypt(secret);
+	const maxAge = 60 * 3;
+
+	cookieStore.set({name:"messageSuccess", value: await encryptedMessage, httpOnly: true, secure: process.env.RUNTIME_ENVIRONMENT === "production", maxAge })
 }