+++ src/app/contact/formProcess/messageSuccess.tsx
@@ -1,21 +1,29 @@
 "use client"
 
-import { useRef } from "react";
+import { useEffect, useRef } from "react";
 import { MediaMap } from "@/utils/mediaMap";
 import { useRouter } from "next/navigation";
 import RpgBorder from "../components/rpgBorder";
 
 export default function MessageSuccess() {
 	const inputRef = useRef<HTMLInputElement>(null);
-	const confirmationSound = new Audio(MediaMap.CONFIRM_SOUND);
+	const confirmationSound = useRef<HTMLAudioElement>(new Audio(MediaMap.CONFIRM_SOUND));
 	const router = useRouter(); 
+
+	useEffect(() => {
+		confirmationSound.current.currentTime = 0;
+		confirmationSound.current
+			.play()
+			.catch(e => console.warn("Audio playback failed: ", e));
+
+	},[]);
 
 	return (
 		<RpgBorder onClick={ () => inputRef.current?.focus() }>
 			<p>Your message has been sent. You will be contacted soon!</p>
 			<button onClick={() => {
-				confirmationSound.currentTime = 0;
-				confirmationSound
+				confirmationSound.current.currentTime = 0;
+				confirmationSound.current
 					.play()
 					.catch(e => console.warn("Audio playback failed: ", e));